src/requirements.txt
Django==3.2.7
psycopg2==2.8.6
django-seed==0.2.2
celery==5.1.2
redis==3.5.3
django_extensions==3.1.3
faker==8.13.2
ipython==7.27.0
black==21.9b0
coverage==5.5
selenium==3.7.0
ipdb==0.10.3
django-selenium-login==2.0.0
slackclient==2.9.3